Solving One Marker Issue
When there is only one active flight in the ACARS map, the map is getting full zoomed to the one point that exists. I have updated it so if there is just one point (google map bound) the map is getting fitted with an offset. The offset is defined in line #202.