Update sidebar_settings.php
Removed

Notify for Updates
Google API Key (for Google Maps)

The need for Google Maps API key and Google Recaptcha keys is documented in the local.config.php file on installation. Not sure how the Notify for Updates would even work now and the Google API key (for Google Maps) isn't needed in the Admin panel. See local.config.php for references on how/where to enter the needed keys.